Added constructors for Product
